feat: do not allow adding/deleting product quotes to offer; disable deleting offer if in status confirm (SL-5557)

+++ modules/offer/controllers/OfferController.php
@@ -4,6 +4,7 @@ namespace modules\offer\controllers;
 
 use common\models\Currency;
 use common\models\LeadPreferences;
+use modules\offer\src\entities\offer\OfferStatus;
 use modules\offer\src\entities\offerProduct\OfferProduct;
 use modules\offer\src\services\OfferService;
 use sales\model\clientChat\socket\ClientChatSocketCommands;
@@ -166,7 +167,7 @@ class OfferController extends FController
                 throw new Exception('Lead (' . $leadId . ') not found', 4);
             }
 
-            $offers = Offer::find()->where(['of_lead_id' => $lead->id])->orderBy(['of_id' => SORT_DESC])->all();
+            $offers = Offer::find()->where(['of_lead_id' => $lead->id])->exceptStatuses([OfferStatus::CONFIRM])->orderBy(['of_id' => SORT_DESC])->all();
 
             if ($offers) {
                 foreach ($offers as $offer) {
@@ -250,6 +251,9 @@ class OfferController extends FController
 
         try {
             $model = $this->findModel($id);
+            if ($model->isConfirm()) {
+                throw new \DomainException('Offer in the "Confirm" status cannot be deleted');
+            }
             if (!$model->delete()) {
                 throw new Exception('Offer (' . $id . ') not deleted', 2);
             }
